Signal shutdown event by closing a `chan struct{}`
Previously a bool channel was used but there is no merit to use
concrete type of channel for the purpose.
In addition to that, I cannot use the method that sends dummy
values to notify child goroutines because the number of children
must be known beforehand; there should be cases it cannot be
known.